Rewrite exhaustion handling
