+++ scripts/modules/publicAPI.mjs
@@ -263,110 +263,3 @@ export class PublicAPI {
     tokens.forEach(token => token.control({releaseOthers: false}));
   }
 }
-
-export class ExhaustionHandler {
-  // Increase exhaustion.
-  static async increaseExhaustion(actor) {
-    if (!(actor instanceof Actor)) {
-      ui.notifications.warn("Invalid actor provided.");
-      return null;
-    }
-
-    // Get current exhaustion effect, if any.
-    const exhaustion = actor.effects.find(i => i.statuses.has("exhaustion"));
-
-    // if exhausted, increase the level.
-    if (exhaustion) {
-      const currentLevel = exhaustion.flags[MODULE].exhaustion;
-      return this.updateExhaustion(currentLevel + 1, actor);
-    }
-
-    // if not exhausted, set to 1.
-    if (!exhaustion) return this.updateExhaustion(1, actor);
-  }
-
-  // Decrease exhaustion.
-  static async decreaseExhaustion(actor, suppress = false) {
-    if (!(actor instanceof Actor)) {
-      ui.notifications.warn("Invalid actor provided.");
-      return null;
-    }
-
-    // Get current exhaustion effect, if any.
-    const exhaustion = actor.effects.find(i => i.statuses.has("exhaustion"));
-
-    // if exhausted, decrease the level.
-    if (exhaustion) {
-      const currentLevel = exhaustion.flags[MODULE].exhaustion;
-      return this.updateExhaustion(currentLevel - 1, actor);
-    }
-
-    // if not exhausted, error.
-    if (!suppress) ui.notifications.warn(`${actor.name} was not exhausted.`);
-    return null;
-  }
-
-  // Update or set exhaustion to specific level.
-  static async updateExhaustion(num, actor) {
-    if (!num.between(0, 11)) {
-      ui.notifications.warn("The provided level was not valid.");
-      return null;
-    }
-
-    if (!(actor instanceof Actor)) {
-      ui.notifications.warn("Invalid actor provided.");
-      return null;
-    }
-
-    // Get current exhaustion effect, if any.
-    const exhaustion = actor.effects.find(i => i.statuses.has("exhaustion"));
-
-    // if num===0, remove it.
-    if (num === 0) return exhaustion?.delete();
-
-    // if num===11, remove it and apply dead.
-    if (num === 11) {
-      await exhaustion?.delete();
-      const dead = foundry.utils.deepClone(CONFIG.statusEffects.find(i => {
-        return (i.id === CONFIG.specialStatusEffects.DEFEATED);
-      }));
-      foundry.utils.mergeObject(dead, {
-        statuses: [dead.id],
-        name: game.i18n.localize(dead.name),
-        "flags.core.overlay": true
-      });
-      return actor.createEmbeddedDocuments("ActiveEffect", [dead]);
-    }
-
-    // Otherwise either update or create the exhaustion effect.
-    const data = {
-      name: game.i18n.localize("ZHELL.StatusConditionExhaustion"),
-      statuses: ["exhaustion"],
-      description: `<p>${game.i18n.format("ZHELL.StatusConditionExhaustionDescription", {level: num})}</p>`,
-      "flags.zhell-custom-stuff.exhaustion": num,
-      icon: "icons/skills/wounds/injury-body-pain-gray.webp",
-      changes: [
-        {key: "system.bonuses.abilities.save", mode: CONST.ACTIVE_EFFECT_MODES.ADD, value: "-@attributes.exhaustion"},
-        {key: "system.bonuses.abilities.check", mode: CONST.ACTIVE_EFFECT_MODES.ADD, value: "-@attributes.exhaustion"},
-        {key: "system.bonuses.mwak.attack", mode: CONST.ACTIVE_EFFECT_MODES.ADD, value: "-@attributes.exhaustion"},
-        {key: "system.bonuses.rwak.attack", mode: CONST.ACTIVE_EFFECT_MODES.ADD, value: "-@attributes.exhaustion"},
-        {key: "system.bonuses.msak.attack", mode: CONST.ACTIVE_EFFECT_MODES.ADD, value: "-@attributes.exhaustion"},
-        {key: "system.bonuses.rsak.attack", mode: CONST.ACTIVE_EFFECT_MODES.ADD, value: "-@attributes.exhaustion"},
-        {key: "system.bonuses.spell.dc", mode: CONST.ACTIVE_EFFECT_MODES.ADD, value: "-@attributes.exhaustion"},
-        {key: "system.attributes.exhaustion", mode: CONST.ACTIVE_EFFECT_MODES.OVERRIDE, value: num},
-      ]
-    };
-
-    // If actor has exhaustion, update it to the new level.
-    if (exhaustion) return exhaustion.update(data);
-
-    // If actor not already exhausted, find and apply.
-    return actor.createEmbeddedDocuments("ActiveEffect", [data]);
-  }
-
-  // Reduce exhaustion on a long rest.
-  static async _longRestExhaustionReduction(actor, data) {
-    if (!data.longRest) return;
-    return ExhaustionHandler.decreaseExhaustion(actor, true);
-  }
-}
